source/OctoBot-Services/requirements.txt
# Services
# Reddit
asyncpraw==7.8.1
# Telegram
python-telegram-bot==21.10   # update alongside supabase for httpx requirement
telethon==1.24.0
# Twitter (associated tentacles are disabled as starting from feb 9 2023, API is now paid only).
# see https://twitter.com/TwitterDev/status/1621026986784337922
# Python-Twitter==3.5
# Google
simplifiedpytrends>=1.1.2
# Ngrok
# 7.2.0 creates an "access denied" error when python tries to communicate with the bin
pyngrok==7.2.2
# chat gpt
openai==1.59.3
# Web
## http server
flask==3.1.0
werkzeug == 3.1.3
# Flask templates
jinja2==3.1.5
## flask minification
flask-compress==1.17
## flask cache
flask-caching==2.3.0
## flask user authentication management
flask-login==0.6.3
## flask CORS management
flask-cors==5.0.0
## user form validators
WTForms==3.2.1
Flask-WTF==1.2.2
## websockets
### used by the webhook service and flask-socketio for the web interface
gevent==24.11.1
### used by flask-socketio with gevent (listed here because multiple libs are usable, force this one)
gevent-websocket==0.10.1
flask-socketio==5.5.0
python-socketio>=5.12.0 # required by flask-socketio >= 5.5.0

# Analysis tools
vaderSentiment==3.3.2
